requirements.txt
selenium
python-dotenv
mysql-connector
